Create qupath.javax-conventions
This moves towards removing Gradle plugins from our version catalog, since they are included within buildSrc instead.
This will require a change to the build scripts in extensions - with the plugin specified directly, and the version included in 'settings.gradle' (as is currently done for the JavaCPP plugin).
