+++ qupath-extension-bioformats/build.gradle.kts
@@ -1,9 +1,9 @@
 plugins {
   id("qupath.extension-conventions")
   id("qupath.publishing-conventions")
+  id("qupath.javafx-conventions")
   `java-library`
 
-  alias(libs.plugins.javafx)
 }
 
 extra["moduleName"] = "qupath.extension.bioformats"
